JAVA  – conditional Statement
Write a Program to check whether the number is a positive or negative
number.
Write a Java program to find the number of days in a month.
Check whether a number is divisible by 5 and 11
WAP to enter the days of the week and print name of the day-- switch
case, Ladderif
WAP to count the total number of 500,100,50,20 and 10 notes in a given
amount
Write a Java program to compare two numbers.